[rescue] Refactor rescue module
1. Refactor the rescue module to not be tied to the uart and xmodem as
   the wire protocol.
2. Implement a `rescue_xmodem` module which provides the xmodem-based
   rescue protocol.

(cherry picked from commit d430519581de9fb254dcdec445bf90e3a21c1c58)
