+++ sw/device/silicon_creator/rom_ext/e2e/rescue/BUILD
@@ -90,6 +90,7 @@ opentitan_test(
         },
         test_cmd = """
             --exec="transport init"
+            --exec="fpga clear-bitstream"
             --exec="fpga load-bitstream {bitstream}"
             --exec="bootstrap --clear-uart=true {firmware}"
             # Set next slot via the rescue protocol
@@ -100,6 +101,7 @@ opentitan_test(
             --exec="gpio apply RESET"
             --exec="gpio remove RESET"
             --exec="console --non-interactive --exit-success='bl0_slot = AA__\r\n' --exit-failure='{exit_failure}'"
+            --exec="fpga clear-bitstream"
             no-op
         """,
     ),
@@ -118,6 +120,7 @@ opentitan_test(
         },
         test_cmd = """
             --exec="transport init"
+            --exec="fpga clear-bitstream"
             --exec="fpga load-bitstream {bitstream}"
             --exec="bootstrap --clear-uart=true {firmware}"
             # Set primary slot via the rescue protocol
@@ -132,6 +135,7 @@ opentitan_test(
             --exec="rescue boot-svc set-next-bl0-slot --primary=SlotA"
             # Check for firmware execution in slot A
             --exec="console --non-interactive --exit-success='bl0_slot = AA__\r\n' --exit-failure='{exit_failure}'"
+            --exec="fpga clear-bitstream"
             no-op
         """,
     ),
